Fix decoding nil ptr.
